requirements.txt
openai
fastapi
python-dotenv
